refactor: 이의신청 컴포넌트로 변경
