+++ src/app/(afterLogin)/workspace/[workspaceId]/workspaceConfirmation/_components/ConfirmationCompo.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image';
 
 import { IWorkoutConfirmationPageProps } from '@/types/workoutConfirmation';
 import ConfirmationProfile from './ConfirmationProfile';
-import confirmationIsObjectionArr from '@/../public/svgs/workspace/workspaceConfirmaion/confirmationIsObjectionArr.svg';
+import ConfirmationCompoObjection from './ConfirmationCompoObjection';
 
 interface IConfirmationCompoProps {
   workoutConfirmationPage: IWorkoutConfirmationPageProps;
@@ -14,18 +14,6 @@ export default function ConfirmationCompo({
   workoutConfirmationPage,
   workspaceId,
 }: IConfirmationCompoProps) {
-  const handleObjection = () => {
-    const targetElement = document.getElementById(
-      `confirmation-${workoutConfirmationPage.objectionId}`
-    );
-
-    if (targetElement) {
-      targetElement.scrollIntoView({
-        behavior: 'auto',
-        block: 'center',
-      });
-    }
-  };
   return (
     <div className='my-10'>
       <ConfirmationProfile workoutConfirmationPage={workoutConfirmationPage} />
@@ -35,26 +23,9 @@ export default function ConfirmationCompo({
         } ml-10 mt-1 flex `}
       >
         {workoutConfirmationPage.isObjection ? (
-          <div
-            onClick={handleObjection}
-            className={`${
-              workoutConfirmationPage.isMine ? 'bg-[#FFEDA6]' : 'bg-[#FDFDFD]'
-            } w-56 h-full py-2 px-2 rounded-lg drop-shadow-lg`}
-          >
-            <div className='flex px-3 pb-1'>
-              <Image
-                src={confirmationIsObjectionArr}
-                alt='confirmationIsObjectionArr'
-              />
-              <span className='text-[#D1D5DB] text-[10px] pl-2'>
-                {workoutConfirmationPage.nickname}님의 인증으로 이동
-              </span>
-            </div>
-            <span className='text-sm'>
-              {workoutConfirmationPage.nickname}님의 운동인증이 이의신청
-              되었어요.
-            </span>
-          </div>
+          <ConfirmationCompoObjection
+            workoutConfirmationPage={workoutConfirmationPage}
+          />
         ) : (
           <div id={`confirmation-${workoutConfirmationPage.objectionId}`}>
             <Link
